"post address"

+++ src/main/java/com/backend/repository/AddressRepository.java
@@ -3,5 +3,6 @@ package com.backend.repository;
 import com.backend.entity.Address;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AddressRepository extends JpaRepository<Address,Integer> {
+public interface AddressRepository extends JpaRepository<Address,Long> {
+
 }
